# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Common Ground API
  description: |
    API for a neighbourhood community platform connecting people through clubs, events, and meals.
    Built as part of **COM5409 — Web Design & Programming**.
  version: 1.0.0
servers:
  - url: http://localhost:3005/api
    description: Local development server
...
paths:
...
  /meals:
    get:
      tags: [Meals]
      summary: Get all meals
      description: Returns all 25 home-cooked meals available
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"

    post:
      tags: [Meals]
      summary: Share a meal
      description: Add a new meal to share with neighbours
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cook
                - meal
                - dietary
                - portions
                - pickupTime
                - location
              properties:
                cook:
                  type: string
                  example: "Maria Rossi"
                meal:
                  type: string
                  example: "Classic Lasagne"
                dietary:
                  type: string
                  enum: [Halal, Vegetarian, Vegan, Healthy]
                  example: "Vegetarian"
                portions:
                  type: integer
                  example: 4
                pickupTime:
                  type: string
                  example: "6:00 PM"
                location:
                  type: string
                  example: "Community Centre"
      responses:
        "200":
          description: Meal shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "401":
          description: Unauthorized
  /meals/{id}/request:
    post:
      tags: [Meals]
      summary: Request a meal
      description: Request a home-cooked meal from a neighbour
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Meal ID
      responses:
        "200":
          description: Requested successfully
        "401":
          description: Unauthorized

  # ============ USERS ============
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      description: Returns the authenticated user's profile information
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

    put:
      tags: [Users]
      summary: Update user profile
      description: Update the authenticated user's profile information
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                phone:
                  type: string
                ageGroup:
                  type: string
                interests:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Profile updated
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: "user@example.com"
        fullName:
          type: string
          example: "John Doe"
        phone:
          type: string
          example: "07700 900123"
        ageGroup:
          type: string
          example: "Adult (40-60)"
        interests:
          type: array
          items:
            type: string
          example: ["Books", "Walking", "Gardening"]
        joinedEvents:
          type: array
          items:
            type: integer
        requestedMeals:
          type: array
          items:
            type: integer
        joinedClubs:
          type: array
          items:
            type: integer
...
    Meal:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cook:
          type: string
          example: "Maria Rossi"
        meal:
          type: string
          example: "Classic Lasagne"
        dietary:
          type: string
          enum: [Halal, Vegetarian, Vegan, Healthy]
          example: "Vegetarian"
        portions:
          type: integer
          example: 4
        pickupTime:
          type: string
          example: "6:00 PM"
        location:
          type: string
          example: "Community Centre"